fixed issue, removing print statements
